Move getAxis implementation to graphUtil
